+++ src/main/java/org/greports/annotations/ColumnGetter.java
@@ -98,6 +98,13 @@ public @interface ColumnGetter {
     int columnWidth() default 1;
 
     /**
+     * The value of this attribute allows to enable the translation
+     * of the cell content.
+     * @return {@code boolean}
+     */
+    boolean translate() default false;
+
+    /**
      * Defines several {@link ColumnGetter} annotations on the same element.
      * @see ColumnGetter
      */