+++ src/main/java/org/greports/converters/AbstractValueConverter.java
@@ -2,5 +2,6 @@ package org.greports.converters;
 
 public interface AbstractValueConverter {
     Object convert(final Object input, final String[] params);
+    @Deprecated
     Class<?> getToClass();
 }