+++ src/main/java/org/greports/utils/ConverterUtils.java
@@ -18,15 +18,4 @@ public class ConverterUtils {
         final AbstractValueConverter valueConverter = ReflectionUtils.newInstance(clazz);
         return valueConverter.convert(value, converter.params());
     }
-
-    public static Object convertValue(final Object value, final List<Converter> converters, final Class<?> toClazz) throws ReportEngineReflectionException {
-        for (final Converter converter : converters) {
-            final Class<? extends AbstractValueConverter> clazz = converter.converterClass();
-            final AbstractValueConverter valueConverter = ReflectionUtils.newInstance(clazz);
-            if(valueConverter.getToClass().equals(toClazz)){
-                return valueConverter.convert(value, converter.params());
-            }
-        }
-        return value;
-    }
 }