+++ CHANGELOG.md
@@ -1,12 +1,22 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [3.2.3] - yyyy-mm-dd
+
+### Added
+- Data cell value translation.
+- Javadoc for some classes.
+### Changed
+- The translations folder moved to the test package.
+### Deprecated
+- The method *AbstractValueConverter.getToClass* is no longer needed. Will be removed in version 4.0
+
 ## [3.2.2] - 2021-03-23
 
 ### Added
-- Javadoc for some classes
+- Javadoc for some classes.
 ### Changed
 - Changed Apache POI version in MANIFEST.MF
-- Some small code improvements
+- Some small code improvements.
 ### Fixed
-- Fixed the obtaining of the result changer
+- Fixed the obtaining of the result changer.